Introduce a new wat to do path resolution in eBPF
The main difference is that we get the path in the correct order inside the kernel
and thus we can do path filtering inside the kernel a bit more easier. We also
removed all assembly code from path resolution. This makes the code easier to understand
and fix if needed.

Regarding the limitations. With the old code we had:
1. Up to 2 mount point with 11 path components each of them (kernels <= 5.3)
2. Up to 32 mount points and 32 path components for each of them (kernels > 5.4)

Now we have:
1. Up to 10 path components independently of mount points (i.e. we can have from 1 up to 10 mount points)  (kernels <= 5.3)
2. Up to 64 path components independently of mount points (kernels > 5.4)

Signed-off-by: Anastasios Papagiannis <tasos.papagiannnis@gmail.com>